Move Collections modules into one folder and start styling collections
