+++ src/client/Components/Collections/CurrentCollection/CurrentCollection.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const CurrentCollection = (props) => {};
+
+export default CurrentCollection;
