+++ src/client/Components/Collections/ShowCollection/ShowCollection.scss
@@ -1,0 +1,7 @@
+.collection-list-container {
+    color: var(--primary-color);
+    background-color: (--lightGray-color);
+    // .collection-title {
+        
+    // }
+}
