--- src/client/Components/CreateCollection/CreateCollection.scss
@@ -1,11 +1,0 @@
-#collectionInput {
-  display: flex;
-  justify-content: center;
-  flex-direction: row;
-  align-items: center;
-}
-
-#addCollectionBtn span {
-  font-weight: bolder;
-  font-size: 1.2rem;
-}
