--- src/client/Components/CurrentCollection/CurrentCollection.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const CurrentCollection = (props) => {
-
-};
-
-export default CurrentCollection;
