+++ src/client/Components/Navbar/Navbar.scss
@@ -1,7 +1,7 @@
 .nav {
   background-color: var(--green-bg);
   min-height: 3em;
-  min-width: 350px;
+  min-width: 300px;
   padding: 1em;
   display: flex;
   flex-direction: row;
