+++ src/client/styles.scss
@@ -13,7 +13,7 @@ body {
     background-color: var(--secondary-color);
     border: 2px solid var(--lightGray-color);
     
-    font-size: 1.25em;
+    font-size: 1.2em;
     font-weight: 700;
     padding: 0.5em 1.25em;
     border-radius: 2em;
